Ensure proper frame timing, interpolate position during rendering
VSYNC: There is an issue on MacOS / Metal (or in the interaction between the graphics backend and SDL2 / this code) regarding vsync that leads to the present() call only blocking on vsync every second time.
Workaround: Make sure to skip physics update in case not enough time has passed, derive vsync timing and use in interpolation.

Timing seems robust now, with the drawback of having two render passes per (displayed) frame. If performance becomes a problem in the future, there is the option to render to a texture instead of the canvas and display it twice.